+++ web-react/src/pages/services/defaulters/DefaulterInfoCard.jsx
@@ -9,7 +9,7 @@ const DefaulterInfoCard = ({ defaulter }) => {
     <Card className="text-center" onClick={() => navigate(defaulter.link)}>
       <Card.Header>{defaulter.title}</Card.Header>
       <PlaceholderCustom
-        loading={defaulter.data === undefined}
+        loading={defaulter.data === undefined || defaulter.data === null}
         className={`fw-bold large-number pb-3 ${defaulter.color}`}
       >
         <Card.Body className={`fw-bold large-number ${defaulter.color}`}>